+++ test/d_count_transfer_manager.js
@@ -201,13 +201,11 @@ contract("CountTransferManager", async (accounts) => {
         it("Should Buy the tokens", async () => {
             // Add the Investor in to the whitelist
 
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor1,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(10))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 500000
@@ -232,13 +230,11 @@ contract("CountTransferManager", async (accounts) => {
         it("Should Buy some more tokens", async () => {
             // Add the Investor in to the whitelist
 
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor2,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(10))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 500000
@@ -260,13 +256,11 @@ contract("CountTransferManager", async (accounts) => {
         it("Should able to buy some more tokens (more than 2 hoders) -- because CountTransferManager is paused", async () => {
             await I_CountTransferManager.pause({ from: account_issuer });
             let snapId = await takeSnapshot();
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor3,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(10))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 500000
@@ -286,13 +280,11 @@ contract("CountTransferManager", async (accounts) => {
         it("Should fail to buy some more tokens (more than 2 holders)", async () => {
             await I_CountTransferManager.unpause({ from: account_issuer });
             // Add the Investor in to the whitelist
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor3,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(10))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 500000
@@ -375,52 +367,44 @@ contract("CountTransferManager", async (accounts) => {
             });
 
             it("add 3 holders to the token", async () => {
-                await I_GeneralTransferManager2.modifyWhitelist(
+                await I_GeneralTransferManager2.modifyKYCData(
                     account_investor1,
                     currentTime,
                     currentTime,
                     currentTime.add(new BN(duration.days(10))),
-                    true,
-                    false,
                     {
                         from: account_issuer,
                         gas: 500000
                     }
                 );
 
-                await I_GeneralTransferManager2.modifyWhitelist(
+                await I_GeneralTransferManager2.modifyKYCData(
                     account_investor2,
                     currentTime,
                     currentTime,
                     currentTime.add(new BN(duration.days(10))),
-                    true,
-                    false,
                     {
                         from: account_issuer,
                         gas: 500000
                     }
                 );
 
-                await I_GeneralTransferManager2.modifyWhitelist(
+                await I_GeneralTransferManager2.modifyKYCData(
                     account_investor3,
                     currentTime,
                     currentTime,
                     currentTime.add(new BN(duration.days(10))),
-                    true,
-                    false,
                     {
                         from: account_issuer,
                         gas: 500000
                     }
                 );
 
-                await I_GeneralTransferManager2.modifyWhitelist(
+                await I_GeneralTransferManager2.modifyKYCData(
                     account_investor4,
                     currentTime,
                     currentTime,
                     currentTime.add(new BN(duration.days(10))),
-                    true,
-                    false,
                     {
                         from: account_issuer,
                         gas: 500000
@@ -513,11 +497,11 @@ contract("CountTransferManager", async (accounts) => {
 
             it("Should allow add a new token holder while transfer all the tokens at one go", async () => {
                 let amount = await I_SecurityToken2.balanceOf(account_investor2);
-                let investorCount = await I_SecurityToken2.getInvestorCount({ from: account_investor2 });
+                let investorCount = await I_SecurityToken2.holderCount({ from: account_investor2 });
                 console.log("current investor count is " + investorCount);
                 await I_SecurityToken2.transfer(account_investor4, amount, { from: account_investor2 });
                 assert((await I_SecurityToken2.balanceOf(account_investor4)).toString(), amount.toString(), { from: account_investor2 });
-                assert(await I_SecurityToken2.getInvestorCount({ from: account_investor2 }), investorCount);
+                assert(await I_SecurityToken2.holderCount({ from: account_investor2 }), investorCount);
             });
         });