+++ test/j_manual_approval_transfer_manager.js
@@ -170,13 +170,11 @@ contract("ManualApprovalTransferManager", accounts => {
         it("Should Buy the tokens", async () => {
             // Add the Investor in to the whitelist
 
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor1,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(30))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 6000000
@@ -201,13 +199,11 @@ contract("ManualApprovalTransferManager", accounts => {
         it("Should Buy some more tokens", async () => {
             // Add the Investor in to the whitelist
 
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor2,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(30))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 6000000
@@ -291,13 +287,11 @@ contract("ManualApprovalTransferManager", accounts => {
         });
 
         it("Add a new token holder", async () => {
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor3,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(10))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 6000000