+++ test/m_presale_sto.js
@@ -241,7 +241,7 @@ contract("PreSaleSTO", async (accounts) => {
             expiryTime = toTime + duration.days(100);
 
             // Add the Investor in to the whitelist
-            let tx = await I_GeneralTransferManager.modifyWhitelist(account_investor1, fromTime, toTime, expiryTime, true, false, {
+            let tx = await I_GeneralTransferManager.modifyKYCData(account_investor1, fromTime, toTime, expiryTime, {
                 from: account_issuer,
                 gas: 6000000
             });
@@ -282,7 +282,7 @@ contract("PreSaleSTO", async (accounts) => {
             expiryTime = toTime + duration.days(100);
 
             // Add the Investor in to the whitelist
-            let tx1 = await I_GeneralTransferManager.modifyWhitelist(account_investor2, fromTime, toTime, expiryTime, true, false, {
+            let tx1 = await I_GeneralTransferManager.modifyKYCData(account_investor2, fromTime, toTime, expiryTime, {
                 from: account_issuer,
                 gas: 6000000
             });
@@ -290,7 +290,7 @@ contract("PreSaleSTO", async (accounts) => {
             assert.equal(tx1.logs[0].args._investor, account_investor2, "Failed in adding the investor in whitelist");
 
             // Add the Investor in to the whitelist
-            let tx2 = await I_GeneralTransferManager.modifyWhitelist(account_investor3, fromTime, toTime, expiryTime, true, false, {
+            let tx2 = await I_GeneralTransferManager.modifyKYCData(account_investor3, fromTime, toTime, expiryTime, {
                 from: account_issuer,
                 gas: 6000000
             });