+++ test/v_tracked_redemptions.js
@@ -187,13 +187,11 @@ contract("TrackedRedemption", async (accounts) => {
         it("Buy some tokens for account_investor1 (1 ETH)", async () => {
             // Add the Investor in to the whitelist
 
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor1,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(30))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 500000
@@ -218,13 +216,11 @@ contract("TrackedRedemption", async (accounts) => {
         it("Buy some tokens for account_investor2 (2 ETH)", async () => {
             // Add the Investor in to the whitelist
 
-            let tx = await I_GeneralTransferManager.modifyWhitelist(
+            let tx = await I_GeneralTransferManager.modifyKYCData(
                 account_investor2,
                 currentTime,
                 currentTime,
                 currentTime.add(new BN(duration.days(30))),
-                true,
-                false,
                 {
                     from: account_issuer,
                     gas: 500000